# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TG Proxy API",
    "description": "API для проксирования манифестов и файлов из GitLab и GitHub репозиториев.\n\nAPI состоит из двух групп эндпоинтов:\n- Публичные эндпоинты: агрегированный манифест, версия каталога, манифесты и файлы проектов, списки версий\n- Админские эндпоинты: управление проектами (CRUD), списки версий проекта, манифест проекта (JSON, с опцией агрегации)\n",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Основной сервер"
    }
  ],
...
  "paths": {
...
    "/{alias}/versions": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "Получить список версий проекта",
        "description": "Возвращает список доступных версий проекта, отсортированный по убыванию",
        "operationId": "getVersions",
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "description": "Алиас проекта",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            },
            "example": "myproject"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное получение списка версий",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "2.0.0",
                    "1.0.25",
                    "1.0.24"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "BasicAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/projects": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Получить список проектов",
        "description": "Возвращает список проектов с пагинацией",
        "operationId": "listProjects",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Количество проектов на странице",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 10
            },
            "example": 10
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Смещение для пагинации",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное получение списка проектов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "BasicAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Создать проект",
        "description": "Создает новый проект с указанными параметрами. URL репозитория автоматически нормализуется (удаляется суффикс `.git`, если присутствует) для обеспечения корректного поиска проектов при обработке зависимостей",
        "operationId": "createProject",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              },
              "example": {
                "alias": "myproject",
                "repo_url": "https://gitlab.com/group/project",
                "token": "glpat-xxxxxxxxxxxxx",
                "description": "Описание проекта",
                "source_name": "gitlab"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Проект успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id"],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "description": "UUID созданного проекта"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "BasicAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Базовая HTTP аутентификация"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT токен в формате Bearer"
      }
    },
    "schemas": {
      "ProjectCreateRequest": {
        "type": "object",
        "required": [
          "alias",
          "repo_url",
          "source_name"
        ],
        "properties": {
          "alias": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Уникальный алиас проекта",
            "example": "myproject"
          },
          "repo_url": {
            "type": "string",
            "format": "uri",
            "description": "Полный URL репозитория",
            "example": "https://gitlab.com/group/project"
          },
          "token": {
            "type": "string",
            "description": "Токен доступа к репозиторию (опционально)",
            "example": "glpat-xxxxxxxxxxxxx"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "description": "Описание проекта (опционально)",
            "example": "Описание проекта"
          },
          "source_name": {
            "type": "string",
            "description": "Имя источника (опционально)",
            "example": "gitlab"
          }
        }
      },
...
      "ProjectResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID проекта",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "alias": {
            "type": "string",
            "description": "Алиас проекта",
            "example": "myproject"
          },
          "repo_url": {
            "type": "string",
            "format": "uri",
            "description": "URL репозитория",
            "example": "https://gitlab.com/group/project"
          },
          "description": {
            "type": "string",
            "description": "Описание проекта",
            "example": "Описание проекта"
          },
          "source_name": {
            "type": "string",
            "description": "Имя источника",
            "example": "gitlab"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Дата создания проекта",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Дата последнего обновления проекта",
            "example": "2024-01-20T14:45:00.000Z"
          }
        }
      },
      "ProjectsListResponse": {
        "type": "object",
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "description": "Список проектов"
          },
          "total": {
            "type": "integer",
            "description": "Общее количество проектов",
            "example": 42
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Сообщение об ошибке",
            "example": "Internal error"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Неверный запрос",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "validation failed: alias is required; repo_url must be a valid URL"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Требуется аутентификация",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Unauthorized"
            }
          }
        }
      },
      "NotFound": {
        "description": "Ресурс не найден",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Project not found"
            }
          }
        }
      },
      "Conflict": {
        "description": "Конфликт (например, проект уже существует)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Project already exists"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Внутренняя ошибка сервера",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Internal error"
            }
          }
        }
      }
    }
  }
}